requirements.txt
flask
gunicorn
pandas
rank_bm25
beautifulsoup4
markdown
python-frontmatter
